Moved concept to indicator mapping to be a class method in AnalysisGraph
